Creating the product slider for the homepage as per issue #13
This commit resolves issue 13 which is also a Sub-Task of #8.
This product slider contains three sections:  "Steal of the day", "Top Sellers" and "Lightning Deals". These 3 sections each contain 5 different products that are the most interesting.  The "Steal of the day" and "Lightning Deals" sections have timers to display the end of the deal.